Optimized function isReadByUninitializedRead
For a given write access, the function now first detects the portion of
the write that overlaps the access represented by the AccessIndex object |ai|.
This way, if a set of writes overwrites only the considered portion, the
function returns false immediately, as the whole part of the write that
overlaps the set is completely overwritten.
Before this optimization, assuming no uninitialized read reading one of
its bytes was detected, the scan of the set continued until either a set
of write accesses had been completely overwritten the considered write (even
the non overlapping portion) or the whole set had been scanned.