Minor fix
On thread start, add the current stack entry (pointed by the stack
pointer) as initialized. This is required because the application's
entry point reads it (through a pop instruction), but it is pushed
before the entry point is executed (and therefore before the tool starts
keeping track of memory accesses). If it is not manually inserted as
initialized, a false positive from the _start function is reported.