Implemented new class + comment extension
MemoryMapParser class has been thought to allow to get the highest address of the
stack address range in an extensible way (memory mapping is available
through different mechanism in different OSs).
Only the LinuxMemoryMapParser has been fully implemented. It opens
/proc/<pid>/maps and parse it to obtain the requested address.

Actually, this class is never used, as it is not required to know which
is the exact beginning of the stack and has been added to the repo in
case it will be required in the future.

Besides, a comment has been extended to be clearer.