Updated testing script + added NullTool rule
Added NullTool rule to makefile.rules to compile an empty pintool, which
just executes the program passed as argument.

Updated test.sh script in order to use smaller random data (1K instead
of 1M) and to fill the execution_times.csv table with further
information, in particular:
[*] The overhead introduced by NullTool w.r.t. original execution;
[*] The overhead introduced by MemTrace w.r.t. NullTool execution;
[*] The overhead introduced by MemTrace w.r.t. original execution;

All the overheads are reported as multiplication factor.