Same instructions are inserted more than once
If required, e.g. the same function reads a local copy of a buffer and
it is called from many call sites, the instruction is reported more than
once with the writes whose content is read by the uninitialized read.

TODO: absolutely clean the code, this is the very first draft of the
correct behaviour, code must be cleaned ASAP