Implemented method compare and a Comparator struct
method compare is similar to operator==, except it does not consider the
execution order of the MemoryAccess objects. It is used in order to
verify if 2 MemoryAccess objects access the same memory area in the same
way, independently from their execution order.

The Comparator class is used as a key_equality function for the map in
function memtrace containing, for each uninitialized read memory access,
the hashes of the contexes it is executed within. In that case, infact,
we are just comparing the memory area they are accessing, not their
execution order, so we can't use the default function operator==.

Also, in order to avoid repeated instructions to cancel themself
(through xor operation) from the context hash, perform rotation before
xoring (so method lrot of the NoOrderHasher class becomes public)