Multiple tracer processes support + fixes
The script launches up to N parallel processes executing the tracer,
where N is the value passed as an argument for -p option.

Fixed problems in case the fuzzer exits immediately (e.g. due to an
error):
[*] If out folder is not created, but tracer_out is created, next time
move only this folder to the backup directory
[*] If the tracer thread is stuck waiting for out folder to be created
and the fuzzer exits without creating it, tracer thread terminates.