Fixed comparson funcs and intervals used for hash
Comparison functions of MemoryAccess objects have been simplified if
they also consider the execution order. Since the order is different for
each stored access, they simply compare the execution order.

The uninitializedInterval has been reintroduced in the computation of
the hash function. The hash function that also considered the execution
order has been simplified to simply return the execution order (or a xor
of the 2 parts composing it if the size of the return type is 32 bit)