Fixed important memory leak
The tool was affected by a quite big memory leak, caused by the fact
that in the instrumentation function Instruction, the opcode and
disassembly arguments to the analysis function were passed as a pointer.
These pointers must remain valid until the end of the execution, because
PIN may cache some of the analysis function instructions in order to
achieve performances, and there's no way to find out when it is safe to
free those pointers.

In order to reduce the problem, the opcode is not passed as a pointer
anymore. PIN defines the OPCODE data type as a 16 bit unsigned integer.
Among the argument types the instrumentation function can use, there's
UINT32, which is a 32 bit unsigned integer. Opcode is now passed as an
UINT32 to avoid allocating memory dynamically, thus avoiding leaks.

This trick can't be done with the disassembly argument, as there's no
way to pass a string to the analysis function except passing a pointer
to it. However, the disassembly string is only meaningful if the DEBUG
flag is enabled, so memory is dynamically allocated only in that case,
while with the DEBUG flag disabled, a NULL pointer is passed.
The analysis function, then, checks whether the passed pointer is NULL
and if it is it simply sets the disassembly to an empty string.