Implemented shadow memory for mmaped malloc+ fixes
In case the size requested to malloc is very high, it simply calls mmap
to allocate memory pages. These pages may be segmented w.r.t. the usual
heap segment, so it is required to allocate a dedicated HeapShadow
object to keep track of them.

Fixed Optional.h to define templated methods, avoiding linking errors.

Added a HeapType class, which is returned by |isHeapAddress|, and
specifies whether an address is a heap address and, in case it is a
mmapped heap address, the index required to retrieve its own HeapShadow
object.