Fixed usage of HeapShadow object shadow memory
Since HeapShadow grows towards high addresses (as the shadow memory
does) we can't know from the first address of the access all the memory
pages we'll need. This may require to add memory pages when the
application is writing to memory.

Note that this is a quite rare situation, where the address starts near
the end of a memory page and continues beyond the start of the next
memory page. If this last page does not have a shadow memory page
allocated, we must allocate it, then, or a segmentation fault will
happen.

Also, allocation of read shadow memory pages have been moved near the
allocation of shadow memory pages (they are perfectly aligned, so they
will need the same number of memory pages allocated)