Avoid using shadow mem to check writes
Instead of checking through the shadow memory if a write access is read
by an uninitialized read access, use the accesses addresses and sizes to
check whether they overlap or not. This saves a lot of time.

For this reason, the readShadow memory is no more used, and it has been,
therefore, removed.

The same happened to all the methods related to the readShadow memory