Fixed problem with writes to be stored
When an uninitialized read access should not be stored again (because it
has already been stored with the same "context") we must store the write
accesses accumulated in any case, otherwise future uninitialized read
accesses may lose their partial inizializations.

Moreover, it is not possible to avoid flushing |lastWriteInstruction|
when we find an uninitialized read (regardless it must be store or not ), otherwise
the "context" of already stored uninitialized reads may result different from the previous one,
even if it is not true.
E.g. store uninitialized read R1 and therefore flush
|lastWriteInstruction| structure, containing all the writes
useful for a following  uninitialized read R2, whose "context" writes
will be empty. Then we find R1 again, which must not be store a second
time. After that, another instance of R2 is found. If we don't flush
|lastWriteInstruction| when we see R1 the second time, the writes that were flushed the first time
because we stored R1, will be now in the structure when we'll find R2 again (let's call it R2'),
and the context for R2' will be different from that of R2. However, the writes performed before R2
and R2' are the same, so we are actually storing R2 twice with the same
"context".