Greatly reduced execution time
Writing all the writes to the main data structure consumed a
lot of time and space, and was unfeasible to run the tool on more complex programs
or even simple programs with bigger inputs.

Instead, permanently store only writes that overlap with uninitialized
reads, and keep the structure |lastWriteInstruction| untouched.
Note that |lastWriteInstruction| only contains the last write access
that accesses a certain address (with a certain size, e.g. instruction
wrote 32B to 0xabcdefff). This structure is NECESSARY in order to allow
the tool to keep track of the writes partially initializing memory read
by uninitialized accesses.