Implemented heuristic to reduce accesses
In order to reduce the number of reported uninitialized read accesses,
an heuristic has been implemented to remove those (not significant)
uninitialized read accesses due to the usage of optimized versions of
the string functions (e.g. strcpy). Those functions, read blocks of 16
or 32 bytes at a time to reduce the number of executed instructions.
This way, however, the functions may load more than the intended string
in SIMD extensions registers, and than re-elaborate their content in
order to return the correct result (e.g. for strcpy, only the bytes
belonging to the string passed as argument are actually copied, the rest
remain in the registers, but are never used in any significant way).

Moreover, during tests resulted that most of the uninitialized reads
reported were due to these functons, and that they were ALL the reported
accesses performed with a size higher than 16 Bytes.

For these reasons, it is better to try and recognize those kind of
accesses, and avoid reporting them, in order to report only the possibly
significant ones.

Of course, it is an heuristics, so it may lead to false negatives (e.g.
in case memcpy is implemented using the same SIMD extension registers).
Moreover, by doing this we are assuming that the ignored uninitialized
reads are not relevant. While this is true for most of them, if a bug
exists (or will exist) in any of those functions implementations, the
tool won't be able to detect it, thus generating another false negative.

However, the number of false negatives of this kind should be very low,
and therefore the benefits of ignoring so many uninitialized accesses
should be higher than the drawbacks.