Reduced reported reads due to free + malloc
In order to avoid that malloc reads bytes considered uninitialized, the
first 16 bytes of a chunk's payload on an x86_64 architecture using
glibc implementation are not re-initialized during a call to free.
This is because free itself writes there some pointers that a following
malloc may read.
Note however that some not relevant uninitialized reads may still be
reported, in some cases.
More infos in issue #2.