Modified binary report structure + fix
A byte has been added into the binary report to specify whether the
considered memory access is a stack memory access or a heap one.

When a write has been found during malloc, and it is not recognized
neither as a stack nor as a heap memory access, we set the currentShadow
to the heap. This is because this happens only during the first write on
the heap, and since before malloc returns we don't know anything about
the heap location, the only way is to temporarily save all writes that
are not on the stack and suppose it is on the heap.