"Re-execute" write accesses after free
After free has been executed and the shadow memory has been
re-initialized, all the write accesses that wrote on the same chunk that
has been freed will be "re-executed", i.e. the memory area they wrote is
set as initialized again.
In order to do this, whenever a heap write access has been detected
during a call to free, it is stored also in the
|mallocTemporaryWriteStorage|, in order to retain a copy of the access
itself.
After the shadow memory has been reset, the tool simply iterates over
all the writes and set the corresponding shadow memory as initialized
again.
Note that the write accesses are not stored again in the main access
map, as they have already been stored.