Better AccessIndex hash function
The new hash function for AccessIndex objects (implemented in AIHasher
functor class) greatly reduces the number of collisions.
Note, however, that there might still be many collisions. In any case,
this is not a cryptographic hash function, so it is not required to have
so few collisions, but the reduced number of collisions also reduces the
buckets sizes, increasing the speed of lookups and iterations.