MemoryAccess default constructor modified
In order to be able and use MemoryAccess objects as elements of a map
and allow us to use operator[] of that map, we are required a default
constructor, which also initialized all the fields of the object.

However, note that everything in the map should be always initialized,
the default constructor is required only to avoid a compiler warning,
which is converted into an error by Intel PIN default Makefile.