Fixed segmentation fault
If the access size was 0 bytes (may happen with some system calls, e.g.
read) function isReadByUninitializedRead raised a segmentation fault,
because it allocated an array of boolean of size 0 and then attempted to
access it to verify whether the write access bytes had been overwritten
by some other write.

In order to fix this, simply check at the beginning if the access size
is 0, and if it is the case, immediately return false.

Indeed, being a write of size 0, it is not possible that any read access
may read its bytes.