Avoid keeping track of 0 sized accesses
Any memory access (both read or write accesses) are not traced if they
have a size of 0 bytes.

This should prevent situations where intervals are computed wrong
(intervals upper bound is computed as accessStart + accessSize - 1. If
accessSize is 0, the upper bound would be lower than the lower bound,
and also overlaps would be computed wrong)