Fixed problem with malloc before entry point
In some cases, heap is allocated before the entry point is started and
calls to malloc and free may read from those allocated memory areas.

If that is the case (cp from Coreutils-6.9.90 and Coreutils-8.31 are
examples) every read performed on the heap areas allocated before the
entry point are considered as uninitialized.

To avoid that, every byte belonging to the heap which is already
allocated at the first malloc call following the entry point is set as
initialized, so that the aforementioned read accesses are not considered
uninitialized anymore