Changed binary overlap parser structure
Instead of directly writing the textual report, the parser simply
returns a queue of tuples in turns containing an AccessIndex object and
a MemoryAccess object.

These objects are quite similar to their C++ versions.

This changes allow to directly call the parser as an executable, in
which case the textual report is generated, or to user the parser as an
external module from within another script.
This will be useful in order to merge all the reports generated during
fuzzing.