Fixed binary report parser
[*] In parse_full_overlaps, removed code to check if a read has
overlapping writes. NOTE: it is highly probable that full overlaps are
going to be removed, and only partial overlaps will be reported (they
are a superset of the full overlaps).
[*] In is_read_by_uninitialized_read, fixed code that checks if a write
is read by any uninitialized read. Instead of checking if it is read by
the uninitialized portion of a read, simply check if it is read by any
portion of the uninitialized read, as we want to report also writes that
partially initialize the read memory area