Removed useless code + added getSize method
Removed all the code required to parse arguments passed through argv, as
it is useless. It was thought to parse some optional arguments which are
not implemented anymore.

Also, the wrapper now accepts a non detemined number of positional
arguments. The first 2 are, in order, the executable path and the input
file path. The others are indices of argv to be set equal to the input
file path.

Finally, a method getSize has been implemented in class ArgBytes so that
it returns the size of the argv array, excluding from the counting the
executable path (argv[0]) and the last NULL element.