Fixed tracer cmd + fixed paths
Before executing the tracer command, arguments are read from the input
file (last 128 bytes) and copied in 2 files, 1 textual (human-readable,
for the user to read) and 1 binary file (to be used in case the tool is
launched again with option --no-fuzzing).
Then, the arguments are appended to the tracer command, also replacing
'@@' with the input file itself.

Fixed paths of the launcher binary and wrapper binary.