Pintool launcher uses execve instead of system
The tool launcher used to merge all the passed arguments into a string,
used to be passed as an argument to function "system".
This way, however, some special characters were misinterpreted (e.g. '!'
was interpreted as an history substitution command).

To solve that, the launcher now creates a NULL terminated array of
string to be used as a new argv array for the process called through
execve