Added __libc_start_main hook usage in memTracer.py
Removed usage of the wrapper program from the launcher (usage of execve
is a problem, as it swaps qemu process away, disabling binary
instrumentation and avoiding the fuzzer to store interesting inputs).

Instead, a library implements a __libc_start_main hook to replace the
original function so that it replaces arguments for the program. The
library is an extension of argvfuzz from AFL++ repository.