Fixed bug discarding some overlapping writes
While looking for write accesses overlapping an uninitialized read, the
program was exiting the while loop as soon as it found an access whose
first accessed byte was higher than the last byte accessed by the read.

Even stopping when we find the first access whose last accessed byte is
higher than the last accessed byte of the read is wrong, because it is
possible that a following access (whose last byte is at an higher
address) partially overlaps with the read.

The only way to be sure to avoid discarding write accesses is to scan
the whole map, starting from a convenient element (first access whose
last byte is higher or equal to the read's first accessed byte)