merge_reports function switched to standalone tool
merge_reports is now a standalone python tool, which can be called
simply passing the tracer output directory path.

However, it is also imported by memTracer.py, so that as soon as the
fuzzing task completes, it can automatically call merge_reports, and
it's not required to do that manually.