Added --ignore option
This option allows the user to specify a list of actual Instruction
Pointers of addresses to be ignored during parsing.

Due to how merge_reports function merges reports, it actually parses all
binary reports as normally, and then filters the result by removing the
instructions to be ignored and all the writes only related to them.

This option is useful when the user has already verified some accesses
and found out that some of them are not relevant. So, he can re-execute
the parser also ignoring the not relevant ones