Better merge_reports function
merge_reports function now checks for equality of MemoryAccess sets by
checking, for each MemoryAccess object, that the offsets of the actualIp
from the base address of the library (or the executable) are the same.

This allows the function to merge also those sets where the MemoryAccess
objects are the same, but they have a different actualIp because the 2
different executions had a different base address for the library.

Of course, the merged report may lose some information of the actual
executions (e.g. the reported actualIp is the actualIp of the first
MemoryAccess set inserted).

However, this way it is easier for a user to understand which inputs
generate the same overlaps, and it is always possible to launch the
single report generator (aka binOverlapParser.py) from inside the folder
containing the input of a particular execution to generate the detailed
report for that execution.
As an alternative, you can launch it from anywhere, passing the folder
path as an argument.