Added bash script to install packages
The script is thought to check for required packages to be installed,
and, if any of those is not detected, an attempt is made to
automatically install it from the apt package manager.

Note that the script will work only on distributions using apt package
manager. In any other case, the user should manually install the
required packages.