Changed merging criterion
Merging is now performed only according to the uninitialized read
access.

In practice, all the sets which has the same instruction as the
uninitialized read are merged together, so that it is easy to check if
the same read reads something written by different instruction (possibly
by different inputs).

Theoretically, this should be an hint that the input to the program
influences the bytes the uninitialized read access is going to read,
simply telling that we can try and manipulate it to perform arbitrary
reads.