Fixed error in map sorter
|lastWriteInstruction| map has been designed to have instructions in
non-decreasing order of the last accessed byte address.

There was an error which caused a replacement of instructions with
different AccessIndex, but whose last accessed bytes were the same.

This also caused some rare errors in the reports.
In order to fix this, simply check if the last accessed byte is the
same.
If it's not, store the instruction with lower address before the one
with higher address.
If they are the same, simply compare the AccessIndex objects and store
them in order.

This allows to replace instructions in the map only if they really have
the same AccessIndex, still keeping the convenient order.