Separate identical access accessing different addr
If a table has 2 identical access sets, but they access different memory
locations, the 2 access sets are considered as different.

This allows to identify arbitrary reads happening with the same set of
instructions.
Usually, this happens when there are specific functions responsible to
write/read memory objects (e.g. think of functions to push/pop objects
from a stack).

With memory location, we mean the couple composed by the offset of the
accessed address from the base address of the section (i.e. STACK or
HEAP base) and the size of the access.