Fixed StackShadow shadow address computation
If the shadow address computation had to ceil the computed shadow
address (there's an ad-hoc boolean variable |needsCeiling| in
|getShadowAddrFromIdx|) there was a problem in case the first computed
address was the address of the last byte of the considered shadow page.

Indeed, by simply increasing its value by 1, it could return the address
of another memory page allocated for something else, thus possibly
causing a segmentation fault, which was even difficult to find, because
it was catched and masked by Intel PIN, which returned a not very clear
error message.

In order to fix that, I changed the signature of method
|getShadowAddrFromIdx|, which now accepts an unsigned* (pointer to an
unsigned), instead of an unsigned for the shadowIdx.
This allows the method itself to modify the value of the passed index,
if needed.

Inside method StackShadow::getShadowAddrFromIdx, infact, if the tool
finds that the computed address points to the last byte of the shadow
page and it requires to be increased, then it increases the value stored
in the passed pointer and returns the address of the first byte of the
next shadow page.