Implement invertBitOrder + fixes
Implemented method invertBitOrder for HeapShadow class, allowing to
reverse a bitsequence in order to make the MSB be on the left and the
LSB on the right, just like it already is with StackShadow and
ShadowRegister classes.

Of course, this requires to fix methods getUninitializedInterval (that
now returns the bit reversed sequence of bits) and computeIntervals,
which is now identical to the homonym method implemented for StackShadow
class.