Small refactoring to avoid code repetition
Method computeIntervals implemented in ShadowBase class and not
overridden by StackShadow and HeapShadow.

This is because we standardized byte position to be (MSB ... LSB).

This is done by using an ad-hoc private method implemented into class
HeapShadow so that it can reverse the bit sequence, and  we can return
(and store) the uninitialized interval with the standardized order.