Mov are not usage + store direct usage
[*] Mov instruction and all its variants are not considered a register
usage anymore. So, associated pending reads won't be reported, but
status is propagated to the destination registers. If and only if any
destination register (or any of its copies) is used, the pending read is
stored
[*] Since mov instructions are not considered as register usage, there's
a special case which must be handled specifically. It is possible
(although improbable) that a mov instruction copies a value from a
register into itself (automov: source and destination registers are the same).
If that is the case, the tool should keep the pending reads associated
to that register, so we can't delete pending read entries associated
with the destination register
[*] If a load is not due to a mov instruction, the read value is
directly used (e.g. add or sub instruction can read one of its operands
directly from memory). In this case, don't leave the read pending, but
directly store it.