Implement register sorter and getAliasingRegisters
Register sorter is a functor class allowing to sort a collection of
shadow register indices (unsigned integer values) sorting them in
decreasing order according to their size in bytes.

The new method getAliasingRegisters is an overloading of the already
existing method which accepts, instead of a REG parameter, a SHDW_REG
parameter.