Avoid consider push and pop as register usages
When the instruction is a push instruction, avoid considering the source
register as used, so that the corresponding uninitialized read is not
stored yet, but the information about uninitialized bytes are
propagated into memory.

Whenever we encounter an uninitialized read executed through a pop
instruction, avoid considering it as a direct usage.
So, just leave the read pending on the destination register so that we
can store that if the register is actually used.