Ignored nop and endbr instructions
For some reason, when endbr instructions are executed, Intel PIN detects
2 explicitly read registers: edi and edx.

By reading Intel x86 Developer manual, however, that instruction should
simply be a marker instruction for those processor having CET feature
enabled, while should be a NOP instruction for those processors not
supporting CET.

Indeed, Intel PIN detects it as a NOP instruction opcode.

Since PIN detects those 2 read registers, if any of them is
uninitialized, it may cause a false positive because it would report the
uninitialized read associated with the register.

To avoid that, both NOP and ENDBR/ENDBR64 instructions are ignored and
not instrumented.

This should not be a problem, because they are instructions that do not
modify the state of memory or of the processor.