Added handlers for direct memory copy and mem init
[*] Direct memory copy: there should be very few instructions allowing
to do that, oen that seems to be frequently used is MOVSD to copy bytes
from a memory location M1 to a memory location M2.

[*] Mem init: there are cases (e.g. system calls) where the memory
location we are writing into should simply be set as initialized. Tha
bypasses a lot of checks and operations by simply setting the whole
memory as initialized and removing the memory range from the
storedPendingUninitializedReads map