Added method setAsInitialized for MemoryAccess
This method allows to reset as initialized a MemoryAccess object that we
previously detected as uninitialized.

This is useful because in some cases we may want to simply propagate
status bits from uninitialized memory to dst registers, but don't want
to report the read access (this is due to a previous store of
uninitialized bytes in the read memory location).

In that case, we just set the read as initialized, so that it won't be
reported, but allows the instruction handler to update the status.