Changed XsaveHandler signatures + fixes
Changed signatures of |getXsaveAnalysisArgs| and
|getXrstorAnalysisArgs|.
Instead of passing a ptr to a CONTEXT object, which is passed wrongly by
Intel PIN, for some reason, retrieve the content of register eax in the
analysis function, and directly pass it as a parameter to the functions.
At this point, CONTEXT object is not needed anymore in these functions,
so we bypassed the problem of the changed pointer.

Fixes:
[*] Fixed the address where XSTATE_BV component is located inside the
xsave area