Fixed non-volatile registers usage + fixed xrstor
When CPUID is used to retrieve information about XSAVE registers,
register rbx, which is non-volatile, is overwritten.
Since the caller expects that register to be untouched on return, the
callee must store it and restore it before returning.

Fixed xrstor behavior by restoring registers which were stored in an
initialized state.