Fixed |copyStoredPendingReads|
The method was not completed. It converted the
storedPendingUninitializedReads entries with the correct entries, but it
never added them back to the map.

Also, it now considers the possibility for the direct memory copy to
also use an uninitialized register as an additional source.
The tags pending on the src register are added to all the new added
ranges.