Changed set<REG> with list<REG> to preserve order
Instead of using sets, we use lists of registers to store src and dst
registers. This allows to preserve the order of the operands, which is
required to handle some instructions precisely.

This should not change performance too much, because insertion and
deletion from the structures is still fast (even faster), and even if we
used set, we required to iterate over all of them most of the time, so
sets give no advantages in those cases.

In some cases, sets were useful to lookup for a certain register, while
now we need to iterate over the whole structure.

This makes the asymptotic complexity of the lookup worse (from O(log(N))
to O(N)). However, src and dst registers inside each structure are so
few that even iterating over all of them can be considered a constant
time complexity operation.