Moved functions + fixed checkDestRegs
Many functions have been moved to an ad-hoc external module

Fix: Now it is possible for an instruction to initialize a destination
register only partially, if it is part of the instruction semantics
(e.g. MOVSD_XMM register propagation).

In order to enable the correct check, the InstructionHandler MUST
register the correct size for the instruction during the |init| method
execution.