Added new emulators + fixed regs list init method
InstructionHandler::handle(list<REG>*) is used by the XsaveHandler when
XRSTOR is called.
Essentially, it restores the state of a register which was stored in an
initialized state.

In that case, we must initialized only the bits related to the passed
registers.

E.g. if YMM1 is uninitialized, but XMM1 is initialized, initializing
XMM1 would initialize also the whole YMM1 register, thus losing
information about uninitialized bytes.
By initializing only bits belonging to XMM1, YMM1 will still result
being uninitialized.