Fixed used register init + update SSE instrs set
Whenever an uninitialized register is used and the associated
uninitialized reads permanently stored, the register is set as
initialized, as it is of no use to propagate information about those
bytes again.
However, in case the initialization of the register also initializes its
super-register, it may discard information about uninitialized bytes.

In order to avoid that, only initialize bytes related to the used
register, leaving those related to ites super-register untouched

Also, we need to know which instructions are part of the SSE ISA
extension. Indeed, in that case, registers XMM0 ~ XMM7 (or XMM31) are
initialized leaving the content of their super-registers (YMM/ZMM
registers) untouched.

In all the other cases, instead, writing register XMMn also overwrites
YMMn and ZMMn.

We called them HybridRegisters, and their behavior depends on the
instruction being executed.

In order to know which opcodes belong to SSE extensions, we simply check
the extension each instruction belongs to, and, if it is an SSE
instruction, we add its OPCODE in a proper set.