Avoid storing/pending initialized indirect reads
During indirect uninitialized reads handling, if a range has some tags
stored in memory, but is initialized, avoid propagating or storing those
tags.

This may happen because, in order to simplify tags storing operations,
all the tags associated to a src register are stored in the written
range, without checking if the tags are actually associated to the part
of the register we are writing.

E.g. if XMM0 is uninitialized in the bit range [32 ~ 127] and we are
storing the first 32 bits through MOVSS, the tag associated with the
uninitialized bytes are stored.

If that memory location is read again, those tags are not stored,
anyway, because the tool checks, for each range which contains tags, if
that range belongs to uninitialized memory.

If it belongs to uninitialized memory (shadow bits unset), then store or
leave the tags pending, otherwise ignore them.