Remove src register+manage stack clash mitigation
In |checkSourceRegisters|, if a register is detected as used (and
therefore the associated uninitialized reads are stored) set the
register itself as initialized and remove its pending reads.
This allows to avoid keeping propagating that same information which is
not useful anymore.

If a program has been compiled with mitigations against stack clash
vulnerability, the probe instructions may be reported as uninitialized
reads.
In order to avoid this, everytime there's a memory read, check if that's
a stack top access.
If it is, check if the previous instruction was a SUB decreasing the
stack pointer by 1 page size.
If this is the case, then ignore the memory read, as it is probably a
probe to avoid stack clash abuse.