Modified stack clash mitigation detection
In order to avoid reporting uninitialized reads executed due to probes added by
the compiler as a mitigation against stack clash, we keep track of each
stack allocation (which is a simple sub of the stack pointer).
After a stack allocation, if anything else is executed in between, the
first uninitialized read that happens inside the newly allocated stack
space is considered as a probe, and therefore not reported.