Fixed stack clash mitigation access check
The boundaries of the check performed for stack accesses (to verify
whether they are part of a stack clash mitigation or not) were wrong.

Stack grows from higher towards lower addresses. Given the start address
and the size of a stack allocation, the end address is given by the
SUBTRACTION of the size from the start address.

Also, this means that the access is performed within the boundaries of the last stack
access if the accessed address is BELOW the start address and ABOVE the
end address.