Fixed stack clash mitigation heuristic
Whenever the stack pointer is decreased by using a register as a second
operand, we assume it is required to probe it, as it is a dynamic
allocation executed through a call to alloca (which is sometimes inlined
as a loop allocating N times a whole page size and the a tail allocation
at the end) or it is the tail allocation of a very big stack allocation
(more than a few pages).

Of these 2, the first situation is the most common.

However, in those cases, if the size we are dynamically allocating is
lower than a page size, the stack probe is performed even if only the
tail allocation is executed.

In that case, if no stack decreasing is performed using the page size as
an immediate, the static boolean variable |stackClashMaybeEnabled| is
still set to false, and therefore the uninitialized read performed by
the probe is reported.

In some binaries (e.g. tail from coreutils) this may generate many false
positives, which may hide something more interesting.

So, we simply consider any stack allocation performed by using a
register as it needs a probe. If something that may be a probe is
executed after that, we ignore it.

Of course, this heuristic may generate some false negative.
However, this should be quite a rare case, comparing to the number of
false positives generated by not applying that.
Moreover, usually stack clash protection is enabled by default by
compilers, so the case where a probe is required should be the most
common case.