Fixed "no-fuzzing" option execution
In case the fuzzer is not executed in parallel (this means it was
already executed and we are running the tracer a second time with the
same inputs) read the arguments from the argv file.

However, we also need to cut the initial part of the input file
generated by the fuzzer, because it contains the raw bytes composing the
argv arguments.