Use private copy of fixed input files
Whenever a program is fuzzed using a fixed input file as an argument (an
input file which is not fuzzed), this is copied inside the testcase
folder, and the path in argv is replaced with the private copy path.

Note that in order to avoid checking ALL the arguments for every
execution, this is done ONLY FOR THE INITIAL ARGUMENTS.

However, this should not be a problem, as it is quite unlikely that the
fuzzer generates a brand new, existing path as an argument.