Fixed path of argv fuzzing marker
Storing the marker inside tracer_out, it won't be available during a
--no-fuzzing run.

Moreover it is a setup valid for all the executions of the tracer, and
it is therefore the same for all of them.

So, store the marker (once) in the root of the tool output (called
fuzz_dir in the script).

This way, it is available even during --no-fuzzing executions.