Fixed explicit src and dst regs detection
By checking if the instruction reads or writes an operand, we are not
considering only the current one, but all the operands.

This may lead to inconsistencies where a register read in order to
compute a memory operand is considered as explicit
(e.g. mov (%rax), %rax: the dst register is rax, and it is an explicit
register. However, INS_RegRContain returns true, because the first
operand uses %rax as a source register to compute the memory address. So
%rax will be added as an explicit src register, even if it is not).

To avoid this wrong behavior, INS_RegWContain and INS_RegRContain have
been replaced by, respectively, INS_OperandWritten and INS_OperandRead,
which check if the current operand is either written or read.